Split types to separate directories
