+++ src/types/Ingredient.ts
@@ -1,0 +1,5 @@
+export type Ingredient = {
+  id?: number;
+  name: string;
+  quantity: number;
+};
