+++ src/types/Meal.ts
@@ -1,0 +1,9 @@
+import { Ingredient } from './Ingredient';
+
+export type Meal = {
+  id?: number;
+  name: string;
+  price: number;
+  availability: boolean;
+  ingredients?: Ingredient[];
+};
