+++ src/types/Order.ts
@@ -1,20 +1,8 @@
+import { Meal } from './Meal';
+
 export type Order = {
   id?: number;
   orderedAt?: string;
   totalPrice: number;
   meals: Meal[];
-};
-
-export type Meal = {
-  id?: number;
-  name: string;
-  price: number;
-  availability: boolean;
-  ingredients?: Ingredient[] ;
-};
-
-export type Ingredient = {
-  id?: number;
-  name: string;
-  quantity: number;
 };
