+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './Client';
+export * from './Ingredient';
+export * from './Order';
+export * from './Meal';
